static_casts and int -> size_t in Arrangement and Boolean Set Operations

+++ Arrangement_on_surface_2/include/CGAL/Arr_point_location/Arr_lm_halton_generator.h
@@ -127,7 +127,7 @@ protected:
     // Create N Halton points. If N was not given to the constructor,
     // set it to be the number of vertices in the arrangement.
     if (num_landmarks == 0)
-      num_landmarks = arr->number_of_vertices();
+      num_landmarks = static_cast<unsigned int>(arr->number_of_vertices());
 
     if (num_landmarks == 0)
       return;
